feat: add Toast common component and update toast color in funding committee status
